Move the code for RepositoryLocationConfig to just be part of DatasetRepositoryConfig, it's not something anyone else should ever use.
